% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{th_browse}
\alias{th_browse}
\title{A shortcut to browse https://opendata.data.go.th/dataset in your browser}
\usage{
th_browse()
}
\description{
A shortcut to browse https://opendata.data.go.th/dataset in your browser
}
