% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{th_get}
\alias{th_get}
\title{Get a dataset from the Open Government Data of Thailand API.}
\usage{
th_get(resource_id, limit = 999999999999, ...)
}
\arguments{
\item{resource_id}{Resource ID of the dataset to download.}

\item{limit}{by default this is set to 999,999,999,999 rows.}
}
\description{
Get a dataset from the Open Government Data of Thailand API.
}
\examples{
th_get("245c98c3-2f92-4d58-a892-3fabf2af0772")
}
