% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{th_package_list}
\alias{th_package_list}
\title{Return a list of the names of the site’s datasets (packages).}
\usage{
th_package_list()
}
\value{
a character vector.
}
\description{
Return a list of the names of the site’s datasets (packages).
}
\examples{
th_package_list()
}
