% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{th_package_search}
\alias{th_package_search}
\title{Searches for packages satisfying a given search criteria.}
\usage{
th_package_search(keyword, as_dataframe = TRUE)
}
\arguments{
\item{keyword}{keyword to search}

\item{as_dataframe}{logical value whether to simplify the result into
a data.frame object. Default as TRUE.}
}
\description{
Searches for packages satisfying a given search criteria.
}
\examples{
# ASCII values means COVID in Thai
th_package_search(keyword = intToUtf8(c(3650L, 3588L, 3623L, 3636L, 3604L)))
}
