% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{th_package_show}
\alias{th_package_show}
\title{Return the metadata of a dataset (package) and its resources.}
\usage{
th_package_show(package_id)
}
\arguments{
\item{package_id}{Package ID.}
}
\value{
a nested list.
}
\description{
Return the metadata of a dataset (package) and its resources.
}
\examples{
th_package_show("3e9d9124-d187-4fc7-b2fb-22c681ceb4fe")
}
