% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thaigov-package.R
\docType{package}
\name{thaigov-package}
\alias{thaigov}
\alias{thaigov-package}
\title{thaigov: Easy Access to the Open Government Data of Thailand}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://asiripanich.github.io/thaigov/}
}

}
\author{
\strong{Maintainer}: Amarin Siripanich \email{asiripanich@icloud.com}

}
\keyword{internal}
